+++ previous-versions/sync-for-commerce-version-1/src/sdk/commercetransactions.ts
@@ -118,7 +118,7 @@ export class CommerceTransactions {
                     );
                 }
                 break;
-            case [401, 404, 409, 429].includes(httpRes?.status):
+            case [401, 402, 403, 404, 409, 429, 500, 503].includes(httpRes?.status):
                 if (utils.matchContentType(contentType, `application/json`)) {
                     res.errorMessage = utils.objectToClass(
                         JSON.parse(decodedRes),
@@ -235,7 +235,7 @@ export class CommerceTransactions {
                     );
                 }
                 break;
-            case [400, 401, 404, 409, 429].includes(httpRes?.status):
+            case [400, 401, 402, 403, 404, 409, 429, 500, 503].includes(httpRes?.status):
                 if (utils.matchContentType(contentType, `application/json`)) {
                     res.errorMessage = utils.objectToClass(
                         JSON.parse(decodedRes),