+++ previous-versions/sync-for-commerce-version-1/src/sdk/models/operations/createconnection.ts
@@ -31,6 +31,12 @@ export class CreateConnectionResponse extends SpeakeasyBase {
     contentType: string;
 
     /**
+     * Your API request was not properly authorized.
+     */
+    @SpeakeasyMetadata()
+    errorMessage?: shared.ErrorMessage;
+
+    /**
      * HTTP response status code for this operation
      */
     @SpeakeasyMetadata()