+++ previous-versions/sync-for-commerce-version-1/src/sdk/models/operations/getcompanydatastatus.ts
@@ -25,7 +25,7 @@ export class GetCompanyDataStatusResponse extends SpeakeasyBase {
      * OK
      */
     @SpeakeasyMetadata({ elemType: shared.DataStatus })
-    dataStatusResponse?: Record<string, shared.DataStatus>;
+    dataStatuses?: Record<string, shared.DataStatus>;
 
     /**
      * Your API request was not properly authorized.