+++ previous-versions/sync-for-commerce-version-1/src/sdk/models/operations/getsyncflowurl.ts
@@ -34,6 +34,12 @@ export class GetSyncFlowUrlResponse extends SpeakeasyBase {
     contentType: string;
 
     /**
+     * The request made is not valid.
+     */
+    @SpeakeasyMetadata()
+    errorMessage?: shared.ErrorMessage;
+
+    /**
      * HTTP response status code for this operation
      */
     @SpeakeasyMetadata()