+++ previous-versions/sync-for-commerce-version-1/src/sdk/models/operations/listcommercelocations.ts
@@ -34,7 +34,7 @@ export class ListCommerceLocationsResponse extends SpeakeasyBase {
     contentType: string;
 
     /**
-     * Your `query` parameter was not correctly formed
+     * Your API request was not properly authorized.
      */
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;