+++ previous-versions/sync-for-commerce-version-1/src/sdk/models/operations/listcompanies.ts
@@ -46,6 +46,12 @@ export class ListCompaniesResponse extends SpeakeasyBase {
     contentType: string;
 
     /**
+     * Your `query` parameter was not correctly formed
+     */
+    @SpeakeasyMetadata()
+    errorMessage?: shared.ErrorMessage;
+
+    /**
      * HTTP response status code for this operation
      */
     @SpeakeasyMetadata()