+++ previous-versions/sync-for-commerce-version-1/src/sdk/models/operations/listconnections.ts
@@ -52,6 +52,12 @@ export class ListConnectionsResponse extends SpeakeasyBase {
     contentType: string;
 
     /**
+     * Your `query` parameter was not correctly formed
+     */
+    @SpeakeasyMetadata()
+    errorMessage?: shared.ErrorMessage;
+
+    /**
      * HTTP response status code for this operation
      */
     @SpeakeasyMetadata()