+++ previous-versions/sync-for-commerce-version-1/src/sdk/models/operations/listintegrations.ts
@@ -40,6 +40,12 @@ export class ListIntegrationsResponse extends SpeakeasyBase {
     contentType: string;
 
     /**
+     * Your `query` parameter was not correctly formed
+     */
+    @SpeakeasyMetadata()
+    errorMessage?: shared.ErrorMessage;
+
+    /**
      * Success
      */
     @SpeakeasyMetadata()