+++ previous-versions/sync-for-commerce-version-1/src/sdk/models/shared/servicecharge.ts
@@ -80,7 +80,7 @@ export class ServiceCharge extends SpeakeasyBase {
     taxes?: ServiceChargeTaxComponentAllocation[];
 
     /**
-     * Total service charge, including taxes.
+     * Total amount of the service charge, including tax.
      */
     @SpeakeasyMetadata()
     @Expose({ name: "totalAmount" })